for #2567, move insert route condition logic form route to optimize
